fix several shields tests after refactor
