+++ components/test/brave_extension/background/api/localeAPI_test.ts
@@ -15,7 +15,7 @@ describe('locale API', () => {
       spy.mockRestore()
     })
     it('calls chrome.i18n.getMessage with the message', () => {
-      localeAPI.getMessage(message)
+      localeAPI.getLocale(message)
       expect(spy).toBeCalledTimes(1)
       expect(spy.mock.calls[0][0]).toBe(message)
     })
